+++ ripper/constants.py
@@ -1,9 +1,9 @@
 ###############################################
 # Constants | Logo and help messages
 ###############################################
-VERSION = 'v2.2.0'
+VERSION = 'v2.3.0'
 USAGE = 'Usage: %prog [options] arg'
-EPILOG = 'Example: dripper -s 192.168.0.1 -p 80 -t 100'
+EPILOG = 'Example: dripper -s tcp://192.168.0.1:80 -t 100'
 
 LOGO_COLOR = f'''[deep_sky_blue1]
 ██████╗ ██████═╗██╗██████╗ ██████╗ ███████╗██████═╗
@@ -16,8 +16,9 @@ LOGO_COLOR = f'''[deep_sky_blue1]
 [grey53]
 It is the end user's responsibility to obey all applicable laws.
 It is just like a server testing script and Your IP is visible.
-
 Please, make sure you are ANONYMOUS!
+
+[u blue link=https://github.com/alexmon1989/russia_ddos]https://github.com/alexmon1989/russia_ddos
 '''
 
 LOGO_NOCOLOR = f'''
@@ -31,8 +32,9 @@ LOGO_NOCOLOR = f'''
 
 It is the end user's responsibility to obey all applicable laws.
 It is just like a server testing script and Your IP is visible.
-
 Please, make sure you are ANONYMOUS!
+
+https://github.com/alexmon1989/russia_ddos
 '''
 
 BANNER = '\n\n[r][deep_sky_blue1]#StandWith[bright_yellow]Ukraine\n'
@@ -75,9 +77,9 @@ ARGS_DEFAULT_PROXY_TYPE = 'socks5'
 
 
 # ==== Defaults ====
-GEOIP_NOT_DEFINED = 'NOT DEFINED'
+GEOIP_NOT_DEFINED = '--'
 CONNECT_TO_HOST_MAX_RETRY = 5
-MIN_SCREEN_WIDTH = 80
+MIN_SCREEN_WIDTH = 90
 MIN_UPDATE_HOST_STATUSES_TIMEOUT = 120
 SUCCESSFUL_CONNECTIONS_CHECK_PERIOD_SEC = 120
 NO_SUCCESSFUL_CONNECTIONS_DIE_PERIOD_SEC = 180